refactor(nordigen): add static mapping structure
No bank seems to use the same Nordigen fields and the rudimentary best
effort guess on which fields to use is not gonna float for all banks.

This adds a structure for mapping every bank independently but defaults
to using or existing mapping to have no breaking changes.
